Restructred Codebase and Cleaned Up Duplicate Files
