+++ .gitignore
@@ -1,2 +1,6 @@
 /node_modules/
+/.vite/
+/dist/
 .env
+.env.local
+.env.*
