--- index.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link rel="icon" type="image/png" href="/favicon.png" />
-    <meta http-equiv="refresh" content="0; url=/index.html">
-</head>
-</html>
