+++ public/about.html
@@ -12,7 +12,7 @@
       rel="stylesheet"
     />
     <script src="https://cdn.tailwindcss.com"></script>
-    <link rel="stylesheet" href="input.css" />
+    <link rel="stylesheet" href="/css/input.css" />
   </head>
   <body class="m-0 p-0 overflow-x-hidden">
     <!-- Header -->
