--- public/auth/signup.html
@@ -1,209 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Signup - Spendwise</title>
-    <link rel="icon" type="image/png" href="/favicon.png" />
-    <script src="https://cdn.tailwindcss.com"></script>
-    <link
-      href="https://fonts.googleapis.com/icon?family=Material+Icons"
-      rel="stylesheet"
-    />
-    <script src="https://accounts.google.com/gsi/client" async defer></script>
-    <link href="https://fonts.googleapis.com/css2?family=Playfair+Display:ital,wght@0,400..900;1,400..900&display=swap" rel="stylesheet">
-    
-    <style>
-      @keyframes fadeUp {
-        from { opacity: 0; transform: translateY(20px); }
-        to { opacity: 1; transform: translateY(0); }
-      }
-      .animate-fade-up {
-        animation: fadeUp 0.6s ease-out forwards;
-      }
-
-      .input-group {
-        position: relative;
-      }
-      .input-field {
-        width: 100%;
-        border: 0;
-        border-bottom: 2px solid #e5e7eb;
-        padding: 0.75rem 0 0.25rem;
-        font-size: 1rem;
-        color: #1f2937;
-        background: transparent;
-        transition: border-color 0.3s ease;
-      }
-      .input-field:focus {
-        border-color: #377f8e;
-        outline: none;
-      }
-      
-      .floating-label {
-        position: absolute;
-        left: 0;
-        top: 0.5rem;
-        font-size: 1rem;
-        color: #9ca3af;
-        transition: all 0.2s ease-out;
-        pointer-events: none;
-      }
-      .input-field:focus + .floating-label,
-      .input-field:not(:placeholder-shown) + .floating-label {
-        top: -0.5rem;
-        font-size: 0.75rem;
-        color: #377f8e;
-        font-weight: 600;
-      }
-
-      .signup-btn {
-        background: linear-gradient(135deg, #377f8e 0%, #2c6571 100%);
-        transition: all 0.3s ease;
-        position: relative;
-        overflow: hidden;
-      }
-      .signup-btn:hover {
-        transform: translateY(-2px);
-        box-shadow: 0 10px 20px -10px rgba(55, 127, 142, 0.5);
-      }
-      .signup-btn:active {
-        transform: scale(0.98);
-      }
-    </style>
-  </head>
-  <body class="bg-gray-50 flex items-center justify-center min-h-screen relative overflow-hidden">
-    
-    <div class="absolute top-[-10%] right-[-10%] w-96 h-96 bg-[#377f8e]/10 rounded-full blur-3xl pointer-events-none"></div>
-    <div class="absolute bottom-[-10%] left-[-10%] w-96 h-96 bg-[#2c6571]/10 rounded-full blur-3xl pointer-events-none"></div>
-
-    <div class="bg-white p-10 rounded-3xl shadow-2xl w-full max-w-md animate-fade-up relative z-10 border border-gray-100">
-      
-      <div class="text-center mb-8">
-        <div class="flex justify-center mb-4">
-            <a href="/" class="transition-transform hover:scale-110 duration-300">
-              <img
-                src="/Images/SpendWise2.png"
-                class="h-16 w-auto"
-                alt="Spendwise Logo"
-              />
-            </a>
-        </div>
-        <h1 class="text-3xl font-bold text-gray-900 font-['Playfair_Display'] mb-2">Create Account</h1>
-        <p class="text-gray-500 text-sm">Join SpendWise to manage your finances</p>
-      </div>
-
-      <form id="signupForm" class="space-y-5">
-        <div class="input-group">
-          <input
-            type="text"
-            id="name"
-            name="name"
-            required
-            class="input-field peer"
-            placeholder=" "
-          />
-          <label for="name" class="floating-label">Full Name</label>
-        </div>
-
-        <div class="input-group">
-          <input
-            type="email"
-            id="email"
-            name="email"
-            required
-            class="input-field peer"
-            placeholder=" "
-          />
-          <label for="email" class="floating-label">Email Address</label>
-        </div>
-
-        <div class="input-group">
-          <input
-            type="password"
-            id="password"
-            name="password"
-            required
-            class="input-field peer pr-10"
-            placeholder=" "
-          />
-          <label for="password" class="floating-label">Password</label>
-          <span
-            class="absolute right-0 top-3 cursor-pointer text-gray-400 hover:text-[#377f8e] transition-colors"
-            onclick="togglePassword('password')"
-          >
-            <i class="material-icons" id="passwordToggleIcon">visibility</i>
-          </span>
-        </div>
-
-        <div class="input-group">
-          <input
-            type="password"
-            id="confirm_password"
-            name="confirm_password"
-            required
-            class="input-field peer pr-10"
-            placeholder=" "
-          />
-          <label for="confirm_password" class="floating-label">Confirm Password</label>
-          <span
-            class="absolute right-0 top-3 cursor-pointer text-gray-400 hover:text-[#377f8e] transition-colors"
-            onclick="togglePassword('confirm_password')"
-          >
-            <i class="material-icons" id="confirmPasswordToggleIcon">visibility</i>
-          </span>
-        </div>
-
-        <button
-          type="submit"
-          class="signup-btn text-white py-3.5 w-full rounded-xl font-bold tracking-wide shadow-md mt-2"
-        >
-          SIGN UP
-        </button>
-
-        <div class="flex items-center my-4">
-            <div class="flex-grow border-t border-gray-200"></div>
-            <span class="px-4 text-xs text-gray-400 uppercase tracking-wider">Or register with</span>
-            <div class="flex-grow border-t border-gray-200"></div>
-        </div>
-
-        <button
-          type="button"
-          id="googleSignup"
-          class="w-full flex items-center justify-center gap-3 py-3 rounded-xl border-2 border-gray-100 hover:border-[#377f8e]/30 hover:bg-gray-50 transition-all duration-300 group"
-        >
-          <img
-            src="https://www.svgrepo.com/show/475656/google-color.svg"
-            class="w-5 h-5 group-hover:scale-110 transition-transform"
-            alt="Google logo"
-          />
-          <span class="text-gray-600 font-medium group-hover:text-[#377f8e]">Sign up with Google</span>
-        </button>
-      </form>
-
-      <p class="text-center text-gray-600 text-sm mt-8">
-        Already have an account?
-        <a
-          href="/auth/login.html"
-          class="text-[#377f8e] font-bold hover:text-[#2c6571] hover:underline transition-all"
-          >Login here</a
-        >
-      </p>
-    </div>
-    <script type="module" src="/js/auth.js"></script>
-    <script>
-      function togglePassword(id) {
-        const input = document.getElementById(id);
-        const iconId =
-          id === "password"
-            ? "passwordToggleIcon"
-            : "confirmPasswordToggleIcon";
-        const icon = document.getElementById(iconId);
-        if (!input || !icon) return;
-        const isHidden = input.type === "password";
-        input.type = isHidden ? "text" : "password";
-        icon.innerText = isHidden ? "visibility_off" : "visibility";
-      }
-    </script>
-  </body>
-</html>
