+++ src/images/main-bg.png
