+++ src/index.html
@@ -13,7 +13,7 @@
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" rel="stylesheet">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css">
     <script src="https://cdn.tailwindcss.com"></script>
-    <link rel="stylesheet" href="./input.css">
+    <link rel="stylesheet" href="/css/input.css">
 </head>
 
 <body class="bg-gray-50">
@@ -28,7 +28,7 @@
         <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
             <div class="flex justify-between h-20">
                 <div class="flex-shrink-0 flex items-center">
-                    <img src="./Images/SpendWise2.png" class="h-20 w-auto" alt="SpendWise">
+                    <img src="/images/SpendWise2.png" class="h-20 w-auto" alt="SpendWise">
                 </div>
                 <div class="hidden md:flex items-center space-x-8">
 
@@ -37,10 +37,10 @@
                     <a href="#features" class="text-[#377f8e] hover:text-black hover:bg-[#f0f0f0] transition-all duration-300 transform hover:scale-105 hover:shadow-md px-4 py-2 rounded-full">Features</a>
                     <a href="#faq" class="text-[#377f8e] hover:text-black hover:bg-[#f0f0f0] transition-all duration-300 transform hover:scale-105 hover:shadow-md px-4 py-2 rounded-full">FAQ</a>
                     <a href="#contact" class="text-[#377f8e] hover:text-black hover:bg-[#f0f0f0] transition-all duration-300 transform hover:scale-105 hover:shadow-md px-4 py-2 rounded-full">Contact</a>
-                    <a href="/auth/login.html"
+                    <a href="/pages/login.html"
                         class="bg-[#377f8e] text-white px-6 py-2 rounded-lg hover:shadow-lg transition-all duration-300">
                         Login</a>
-                    <a href="/auth/signup.html"
+                    <a href="/pages/signup.html"
                         class="bg-[#377f8e] text-white px-6 py-2 rounded-lg hover:shadow-lg transition-all duration-300">
                         Sign Up</a>
                     <button id="logout" class="hidden bg-[#377f8e] text-white px-6 py-2 rounded-lg hover:shadow-lg
@@ -50,7 +50,7 @@
                             <i class="fas fa-sun text-xl" id="theme-toggle-light-icon"></i>
                             <i class="fas fa-moon text-xl hidden" id="theme-toggle-dark-icon"></i>
                         </button>
-                        <a href="./dashboard.html" class="bg-gradient-to-r from-[#377f8e] to-[#2c6571] text-white px-6 py-2 rounded-lg hover:shadow-lg transition-all duration-300 transform hover:scale-105">
+                        <a href="/pages/dashboard.html" class="bg-gradient-to-r from-[#377f8e] to-[#2c6571] text-white px-6 py-2 rounded-lg hover:shadow-lg transition-all duration-300 transform hover:scale-105">
                             Dashboard
                         </a>
                     </div>
@@ -80,7 +80,7 @@
                 class="block px-3 py-2 text-[#377f8e] hover:bg-gray-50 rounded-lg transition-colors duration-200">FAQ</a>
             <a href="#contact"
                 class="block px-3 py-2 text-[#377f8e] hover:bg-gray-50 rounded-lg transition-colors duration-200">Contact</a>
-            <a href="./dashboard.html"
+            <a href="/pages/dashboard.html"
                 class="block px-3 py-2 bg-[#377f8e] text-white rounded-lg text-center hover:bg-[#2c6571] transition-colors duration-200">
                 Dashboard
             </a>
@@ -100,7 +100,7 @@
                     solution.
                 </p>
                 <div class="flex flex-col sm:flex-row gap-4 justify-center md:justify-start">
-                    <a href="./dashboard.html"
+                    <a href="/pages/dashboard.html"
                         class="bg-[#377f8e] text-white px-8 py-3 rounded-lg text-lg hover:bg-[#1f3c42] text-center">
                         Get Started
                     </a>
@@ -111,7 +111,7 @@
                 </div>
             </div>
             <div class="md:w-1/2 mb-8 md:mb-0 fade-in-right p-4">
-                <img src="./Images/main-bg.png" alt="Expense Tracking"
+                <img src="/images/main-bg.png" alt="Expense Tracking"
                     class="w-full h-auto rounded-2xl shadow-2xl hover:transform hover:scale-105 transition-transform duration-500">
             </div>
         </div>
@@ -491,7 +491,7 @@
             <div class="grid grid-cols-1 md:grid-cols-3 gap-8">
                 <!-- Logo and Description -->
                 <div class="text-center md:text-left">
-                    <img src="./Images/SpendWise.png" class="h-16 w-auto mx-auto md:mx-0 mb-4" alt="SpendWise">
+                    <img src="/images/SpendWise.png" class="h-16 w-auto mx-auto md:mx-0 mb-4" alt="SpendWise">
                     <p class="text-gray-200 text-sm">Smart financial management for a better future</p>
                 </div>
 
@@ -532,7 +532,7 @@
         </div>
     </footer>
 
-    <script type="module" src="script.js"></script>
+    <script type="module" src="/js/main.js"></script>
     
     <script>
         let scrollBtn = document.getElementById("scrollBtn");
