--- src/js/protectedRoute.js
@@ -1,16 +1,0 @@
-import { auth } from './firebase.js';
-
-export const checkAuth = (redirectIfUnauthenticated = true) => {
-  return new Promise((resolve) => {
-    auth.onAuthStateChanged((user) => {
-      if (user) {
-        resolve(user);
-      } else {
-        if (redirectIfUnauthenticated) {
-          window.location.href = '/src/auth/login.html';
-        }
-        resolve(null);
-      }
-    });
-  });
-};
