+++ vite.config.js
@@ -1,11 +1,17 @@
 import { defineConfig } from 'vite';
 
 export default defineConfig({
+  root: 'src',
+  publicDir: '../public',
   build: {
+    outDir: '../dist',
+    emptyOutDir: true,
     rollupOptions: {
       input: {
-        main: 'public/index.html',
-        dashboard: 'public/dashboard.html',
+        main: 'index.html',
+        dashboard: 'pages/dashboard.html',
+        login: 'pages/login.html',
+        signup: 'pages/signup.html',
       },
     },
   },
